took table panel into its own file
